forwarder: client: Use cached version of attachments for deleted messages
Whether to use proxy_url rather than url when downloading the attachment.
This will allow attachments to be saved after deletion more often, compared
to the regular URL which is generally deleted right after the message is
deleted. Note that this can still fail to download deleted attachments if
too much time has passed and it does not work on some types of attachments.

Signed-off-by: Koko <github@kokosoftware.pl>